feat(FTDA): Added endpoint for the toggle-complete. fix(FTDA): Add the togglefunction to the TaskDetail component.
